requirements.txt


# Hugging Face
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0

# Tokenizer
tokenizers>=0.15.0
# Data Analysis & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Config
pyyaml>=6.0